Sea: separate psv specific parts from input
