+++ icicle.cabal
@@ -158,7 +158,6 @@ library
                        Icicle.Simulator
 
                        Icicle.Sea.Error
-                       Icicle.Sea.Eval
                        Icicle.Sea.FromAvalanche
                        Icicle.Sea.FromAvalanche.Analysis
                        Icicle.Sea.FromAvalanche.Base
@@ -170,6 +169,12 @@ library
                        Icicle.Sea.Chords
                        Icicle.Sea.Chords.File
                        Icicle.Sea.Chords.Parse
+                       Icicle.Sea.Eval
+                       Icicle.Sea.Eval.Base
+                       Icicle.Sea.Eval.Psv
+                       Icicle.Sea.IO
+                       Icicle.Sea.IO.Base
+                       Icicle.Sea.IO.Base.Input
                        Icicle.Sea.IO.Psv
                        Icicle.Sea.IO.Psv.Base
                        Icicle.Sea.IO.Psv.Input
