+++ src/Icicle/Sea/Eval/Psv.hs
@@ -1,0 +1,21 @@
+{-# LANGUAGE EmptyDataDecls #-}
+{-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE NoImplicitPrelude #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE PatternGuards #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TupleSections #-}
+module Icicle.Sea.Eval.Psv (
+    PsvStats (..)
+  , PsvState
+  ) where
+
+import           P hiding (count)
+
+
+data PsvStats = PsvStats {
+    psvFactsRead    :: Int64
+  , psvEntitiesRead :: Int64
+  } deriving (Eq, Ord, Show)
+
+data PsvState
