+++ src/Icicle/Sea/IO/Base.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+{-# LANGUAGE LambdaCase        #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE PatternGuards     #-}
+{-# LANGUAGE TupleSections     #-}
+module Icicle.Sea.IO.Base (module X) where
+
+import Icicle.Sea.IO.Base.Input as X
